fix some small issues on the replication of data between tables
Former-commit-id: 0238e0e6204c8aab43693dfa03de3d6557306ead

+++ covigator/processor/ena_processor.py
@@ -105,7 +105,7 @@ class EnaProcessor(AbstractProcessor):
             job.num_reads = int(data.numreads.loc[0])
             job.num_reads = int(data.numreads.loc[0])
             job.covered_bases = int(data.covbases.loc[0])
-            job.coverage = int(data.coverage.loc[0])
+            job.coverage = float(data.coverage.loc[0])
         except Exception as e:
             raise CovigatorErrorProcessingCoverageResults(e)
